add connectGithub router
