+++ src/components/DefaultHeader/index.vue
@@ -183,7 +183,7 @@ const goMiwaspace = () => {
 const githubInstall = () => {
   // 关联github
   const state = new Date().getTime();
-  const url = `${oauthUrl.value}?client_id=${clientId.value}&scope=read:user&state=${state}&redirect_uri=${apiUrl.value}/projects/installations`;
+  const url = `${oauthUrl.value}?client_id=${clientId.value}&scope=read:user&state=${state}&redirect_uri=${apiUrl.value}/projects/connectGithub`;
   const myWindow = window.open(url, 'login-github', `modal=yes,toolbar=no,titlebar=no,menuba=no,location=no,top=100,left=500,width=800,height=700`)
   myWindow?.focus();
   // window.close();
